feat: string class work + refactorings

+++ Source/Engine/Source/System/SystemInfo.cpp
@@ -5,6 +5,7 @@
 #include "SystemInfo.h"
 
 #include <array>
+#include <vector>
 #include <Core/System.h>
 #include <Core/Types.h>
 
@@ -113,7 +114,7 @@ void GetCPUInfo(SystemInfo& info)
     } while (current.QuadPart - count.QuadPart < wait.QuadPart);
 
     SetThreadPriority(thread_info, priority);
-    info.cpuMhz = static_cast<f32>(static_cast<f32>((__rdtsc() - start) << 5) / 1000000.0f);
+    info.cpuMhz = static_cast<f32>((__rdtsc() - start) << 5) / 1000000.0f;
 }
 
 void GetMemoryInfo(SystemInfo& info)
